:recycle: Rewrite demo version
